Agregados test de herramienta

+++ src/fiuba/algo3/tp2/herramienta/durabilidad/Durabilidad.java
@@ -1,8 +1,10 @@
 package fiuba.algo3.tp2.herramienta.durabilidad;
 
+import java.math.BigDecimal;
+
 public interface Durabilidad {
 
 	void reducir();
 
-	float getValor();
+	BigDecimal getValor();
 }
