+++ src/fiuba/algo3/tp2/jugador/Inventario.java
@@ -3,6 +3,8 @@ package fiuba.algo3.tp2.jugador;
 import fiuba.algo3.tp2.herramienta.Herramienta;
 import fiuba.algo3.tp2.herramienta.HerramientaFactory;
 
+import java.math.BigDecimal;
+
 public class Inventario {
 
     private Herramienta herramientaInicial;
@@ -11,7 +13,7 @@ public class Inventario {
         this.herramientaInicial = HerramientaFactory.newHachaDeMadera();
     }
 
-    public float getDurabilidadDelHachaDeMadera() {
+    public BigDecimal getDurabilidadDelHachaDeMadera() {
         return this.herramientaInicial.getDurabilidad();
     }
 }
